+++ Tests/MapboxCoreNavigationTests/NavigationServiceTests.swift
@@ -597,7 +597,7 @@ class NavigationServiceTests: XCTestCase {
     }
 
     func testUnimplementedLogging() {
-        unimplementedTestLogs = []
+        _unimplementedLoggingState.clear()
 
         let options =  NavigationRouteOptions(coordinates: [
                    CLLocationCoordinate2D(latitude: 38.853108, longitude: -77.043331),
@@ -618,16 +618,8 @@ class NavigationServiceTests: XCTestCase {
             service.locationManager(locationManager, didUpdateLocations: [location])
         }
 
-        guard let logs = unimplementedTestLogs else {
-            XCTFail("Unable to fetch logs")
-            return
-        }
-
-        let ourLogs = logs.filter { $0.0 == "EmptyNavigationServiceDelegate" }
-
-        XCTAssertEqual(ourLogs.count, 7, "Expected logs to be populated and expected number of messages sent")
-        unimplementedTestLogs = nil
-    }
+        XCTAssertEqual(_unimplementedLoggingState.countWarned(forTypeDescription: "EmptyNavigationServiceDelegate"), 7, "Expected logs to be populated and expected number of messages sent")
+    }    
 }
 
 class EmptyNavigationServiceDelegate: NavigationServiceDelegate {}